refactor(dns): split set.rs into set/ by concern
dns set's reconciliation logic (plan/outcome/write) lived in one
1107-line file. Split into set/{mod,plan,outcome,write}.rs so each
concern has its own file, matching the domain/dns file-per-subcommand
convention.
